Add automatic cleanup (#10)
While working on making a short 2-liner for Alex, I noticed that you get
some nasty-grams printed if you don't do cleanup.

I looked into it a bit and decided we could use atexit handlers to do
the cleanup. This appears to work well and simplifies the examples and
user experience.